+++ ide/RightInfo.php
@@ -1,0 +1,39 @@
+<?php
+declare(strict_types=1);
+
+namespace Win32Service;
+
+final class RightInfo
+{
+
+    public function getFullUsername(): string
+    {
+
+    }
+
+    /** @return array<int, string> */
+    public function getRights(): array
+    {
+
+    }
+
+    public function isGrantAccess(): bool
+    {
+
+    }
+
+    public function isDenyAccess(): bool
+    {
+
+    }
+
+    public function getUsername(): string
+    {
+
+    }
+
+    public function getDomain(): ?string
+    {
+
+    }
+}