+++ win32service_right.h
@@ -1,0 +1,10 @@
+//
+// Created by jbnahan on 14/07/24.
+//
+
+#ifndef WIN32SERVICE_WIN32SERVICE_RIGHT_H
+#define WIN32SERVICE_WIN32SERVICE_RIGHT_H
+
+long change_service_right(char *service, char *username, long right, long type, char *machine, char *returnMessage);
+
+#endif //WIN32SERVICE_WIN32SERVICE_RIGHT_H